Updated app engine configuration to python313 runtime

+++ dev/build-dev.sh
@@ -6,10 +6,12 @@
 # Run this from /frontend directory
 
 # Move files from frontend/dist to dev directories accessed by Flask
-mkdir ../static
-mkdir ../static/img
-# mkdir ../templates
-cp ./dist/*.js ./dist/*.css ./dist/static/img/favicon.ico ../static
-cp ./dist/static/img/* ../static/img
-cp ./dist/index.html ../templates
-
+if [ -f "./dist/index.html" ]; then
+  mkdir -p ../static/img
+  mkdir -p ../templates
+  cp ./dist/*.js ./dist/*.css ./dist/static/img/favicon.ico ../static
+  cp ./dist/static/img/* ../static/img
+  cp ./dist/index.html ../templates
+else
+  echo "Skipping dev build copy: ./dist/index.html not found."
+fi
